Added ability flags and descriptions

+++ include/pokemon.h
@@ -244,7 +244,7 @@ struct BattleMove
     u8 secondaryEffectChance;
     u8 target;
     s8 priority;
-    u8 flags;
+    u16 flags;
 	u8 category:1;	// 0: Physical. 1: Special.
 };
